+++ pysrc/bytewax/outputs.py
@@ -83,7 +83,7 @@ class StatefulSink(ABC):
         a finite input.
 
         This is not guaranteed to be called. It will not be called
-        during a crash.
+        during an abrupt or abort shutdown.
 
         """
         pass
@@ -197,7 +197,7 @@ class StatelessSink(ABC):
         a finite input.
 
         This is not guaranteed to be called. It will not be called
-        during a crash.
+        during an abrupt or abort shutdown.
 
         """
         pass